Add fp32_sincos.h helper
